Fix negative complex number bug. Move check_version to astmonkey.utils.
