+++ astmonkey/utils.py
@@ -1,5 +1,4 @@
 import ast
-
 import sys
 
 
@@ -39,3 +38,9 @@ class CommaWriter:
                 self.write_func(' ')
         else:
             self.write_func(', ')
+
+def check_version(from_inclusive=None, to_exclusive=None):
+    if (not from_inclusive or sys.version_info >= from_inclusive) \
+            and (not to_exclusive or sys.version_info < to_exclusive):
+        return True
+    return False
